init scripts asks before shutdown

+++ init.sh
@@ -9,6 +9,31 @@ echo
 echo installing python and pip
 sudo apt install python3 python3-pip -y
 
+# setting repos environment variables
+
+export DNS_REPOS_ROOT=https://github.com/mafazaa-org/
+export DNS_INIT=dns-init
+export DNS_SERVER=dns-server
+export DNS_CHECK=dns-check
+export DNS_IP_CONFIRM=dns-ip-confirm
+
+echo export DNS_REPOS_ROOT=$DNS_REPOS_ROOT >> ~/.bashrc
+echo export DNS_INIT=$DNS_INIT >> ~/.bashrc
+echo export DNS_SERVER=$DNS_SERVER >> ~/.bashrc
+echo export DNS_CHECK=$DNS_CHECK >> ~/.bashrc
+echo export DNS_IP_CONFIRM=$DNS_IP_CONFIRM >> ~/.bashrc
+
+source ~/.bashrc
+
+crontab_file=/var/spool/cron/crontabs/dns-admin
+
+echo DNS_REPOS_ROOT=$DNS_REPOS_ROOT >> $crontab_file
+echo DNS_INIT=$DNS_INIT >> $crontab_file
+echo DNS_SERVER=$DNS_SERVER >> $crontab_file
+echo DNS_CHECK=$DNS_CHECK >> $crontab_file
+echo DNS_IP_CONFIRM=$DNS_IP_CONFIRM >> $crontab_file
+
+
 ## freeing port 53
 
 echo 
@@ -27,22 +52,10 @@ sudo ln -sf /run/systemd/resolve/resolv.conf /etc/resolv.conf
 echo
 echo congratulations, now we are going to shutdown the system
 echo 
-read
 
-# setting repos environment variables
+echo shutdown now? (y/n)
+read shutdown_now
 
-export DNS_REPOS_ROOT=https://github.com/mafazaa-org/
-export DNS_INIT=dns-init
-export DNS_SERVER=dns-server
-export DNS_CHECK=dns-check
-export DNS_IP_CONFIRM=dns-ip-confirm
-
-echo export DNS_REPOS_ROOT=https://github.com/mafazaa-org/ >> ~/.bashrc
-echo export DNS_INIT=dns-init >> ~/.bashrc
-echo export DNS_SERVER=dns-server >> ~/.bashrc
-echo export DNS_CHECK=dns-check >> ~/.bashrc
-echo export DNS_IP_CONFIRM=dns-ip-confirm >> ~/.bashrc
-
-source ~/.bashrc
-
-sudo shutdown -h now
+if [ shutdown_now == "y" ]; then
+    sudo shutdown -h now
+fi
